backend/requirements.txt
# Backend framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0
pydantic>=2.5.2

# PDF processing
pypdf>=3.17.0

# LangChain (MODERN – Python 3.12 compatible)
langchain-core>=0.1.52
langchain-community>=0.0.38
langchain-text-splitters>=0.0.1
langchain-groq>=0.1.3

# Vector DB
chromadb>=0.4.24

# Embeddings
sentence-transformers>=2.6.1
langchain-huggingface>=0.0.1

